requirements.txt
# 根据 Python 版本自动选择兼容依赖
# Python >= 3.10
pandas==2.2.2; python_version >= "3.10"
numpy==2.1.3; python_version >= "3.10"
scikit-learn==1.5.2; python_version >= "3.10"
networkx==3.3; python_version >= "3.10"
matplotlib==3.9.2; python_version >= "3.10"
seaborn==0.13.2; python_version >= "3.10"
pyyaml==6.0.2
scipy==1.13.1; python_version >= "3.10"

# 3.8 <= Python < 3.10（常见于 Win10）
pandas==1.5.3; python_version >= "3.8" and python_version < "3.10"
numpy==1.23.5; python_version >= "3.8" and python_version < "3.10"
scikit-learn==1.2.2; python_version >= "3.8" and python_version < "3.10"
networkx==3.1; python_version >= "3.8" and python_version < "3.10"
matplotlib==3.7.5; python_version >= "3.8" and python_version < "3.10"
seaborn==0.12.2; python_version >= "3.8" and python_version < "3.10"
scipy==1.10.1; python_version >= "3.8" and python_version < "3.10"

# 3.7 <= Python < 3.8（如旧环境）
pandas==1.3.5; python_version >= "3.7" and python_version < "3.8"
numpy==1.21.6; python_version >= "3.7" and python_version < "3.8"
scikit-learn==1.0.2; python_version >= "3.7" and python_version < "3.8"
networkx==2.6.3; python_version >= "3.7" and python_version < "3.8"
matplotlib==3.5.3; python_version >= "3.7" and python_version < "3.8"
seaborn==0.11.2; python_version >= "3.7" and python_version < "3.8"
scipy==1.7.3; python_version >= "3.7" and python_version < "3.8"
